test(install): consolidate 11 install-domain test files → 3 + shared fixture
The install-domain suite was 11 files each re-declaring the same boilerplate
(temp dir + synthetic AgentConfig + register/cleanup into the agents map),
differing only in a name suffix and useFlatFiles. Merged them into 3 subject
files backed by one fixture helper. Every it()/expect() copied VERBATIM — no
assertion rewritten.

- __tests__/helpers/install-fixture.ts — makeInstallFixture(name, {useFlatFiles?,
  withCommands?}) + listSkillDirs(); owns the repeated setup once
- install-core.test.ts     ← install-all, install-specific, install-commands,
                              install-uninstall, installer-scripts-copy   (31 tests)
- install-profiles.test.ts ← installer-align, installer-thclaws            (19 tests)
- install-e2e.test.ts      ← e2e-install, e2e-features, integration,
                              installer-behavior                           (26 tests)

Files that don't fit the synthetic-agent pattern (installer-scripts-copy's dual
agents, installer-behavior's $-shell setup, integration's real ~/.config paths)
keep their original setup verbatim inside a block scope — helper used only where
it's a clean fit.

Coverage proof: full fast suite was 323 pass with old+new both present; after
removing the 11 originals it is 247 (= 323 − 76), 0 fail. The 76 install-domain
tests are exactly preserved, none lost or gained. __tests__ file count 24 → 13.
